test: fix UT test failure
